cleaned up profile routes
